Fixed 'include ../blah' and other things
This commit fixes a bug in the machine generation of config files where
a .ktb that did an include containing .. would not properly include from
the parent folder. This meant that, for example, all of the keys in al-el
were flagged as unbound.

This commit also adds an extra remapping for al-el's 'Joystick Enter' key to
jcenter. That would usually be something to do manually, but jenter would be
an error.